Move velox_dwio_parquet_field_id to parent folder
