+++ velox/connectors/hive/iceberg/CMakeLists.txt
@@ -29,10 +29,13 @@ velox_add_library(
 velox_link_libraries(
   velox_hive_iceberg_splitreader
   velox_connector
-  velox_dwio_parquet_field_id
   velox_functions_iceberg
   Folly::folly
 )
+
+if(VELOX_ENABLE_PARQUET)
+  velox_link_libraries(velox_hive_iceberg_splitreader velox_dwio_parquet_field_id)
+endif()
 
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(tests)
