+++ velox/dwio/parquet/CMakeLists.txt
@@ -25,6 +25,7 @@ endif()
 
 velox_add_library(velox_dwio_parquet_reader RegisterParquetReader.cpp)
 velox_add_library(velox_dwio_parquet_writer RegisterParquetWriter.cpp)
+velox_add_library(velox_dwio_parquet_field_id INTERFACE ParquetFieldId.h)
 
 if(VELOX_ENABLE_PARQUET)
   velox_link_libraries(velox_dwio_parquet_reader velox_dwio_native_parquet_reader xsimd)
