+++ velox/dwio/parquet/ParquetFieldId.h
@@ -16,6 +16,8 @@
 
 #pragma once
 
+#include <vector>
+
 namespace facebook::velox::parquet {
 /// Parquet field IDs during write operations. Each ID must be unique positive
 /// number, do not need to be sequential.
